feat: use traits for validation error handling and aria-describedby generation
